+++ SRC/Model/DAO/AutoArDAO.php
@@ -1,10 +1,10 @@
 <?php
 $path = dirname(__DIR__, 2);
 include($path . "/Connection/Dbconn.php");
-function AutoTipusTarolo($Marka,$Tipus,$Premium,$Ar)
+function AutoTipusTarolo($Marka, $Tipus, $Ar)
 {
     $kapcsolat = $GLOBALS["kapcsolat"];
-    $sql = "INSERT INTO `autokolcsonzo`.`autotipus` (`ID`, `Márka`,`Tipus`,  `Prémium`, `Ar`) VALUES (NULL, '$Marka','$Tipus', '$Premium', '$Ar')"; // modositott
+    $sql = "INSERT INTO `autokolcsonzo`.`autotipus` (`ID`, `Márka`,`Tipus`, `Ar`) VALUES (NULL, '$Marka','$Tipus', '$Ar')"; // modositott
 
     $üzenet = "Az autotipus felvétel ";
     $ok = mysqli_query($kapcsolat, $sql);
